move reference data and links creation into own services
